Added date and custom list functionality

+++ app.js
@@ -15,23 +15,28 @@ const itemsSchema = new mongoose.Schema({
     name: String
 })
 
+const listSchema = new mongoose.Schema({
+    name: String,
+    items: [itemsSchema]
+})
+
 const Item = mongoose.model("Item", itemsSchema)
+const List = mongoose.model("List", listSchema)
 
-//Sample data for Item model
-// const item1 = new Item({
-//     name: "Welcome to your todolist"
-// })
+//Sample data for List model
+const item1 = new Item({
+    name: "Welcome to your new custom list!"
+})
 
-// const item2 = new Item({
-//     name: "CLick the + button to add a new item."
-// })
+//Get current date
+let today = new Date()
+let options = {
+    weekday: "long",
+    day: "numeric",
+    month: "long"
+}
 
-// const item3 = new Item({
-//     name: "<--- Click this to delete an item."
-// })
-
-// const defaultItems = [item1, item2, item3]
-
+let day = today.toLocaleDateString('en-us', options)
 
 
 //Routes
@@ -39,17 +44,8 @@ const Item = mongoose.model("Item", itemsSchema)
 //Render current list items
 app.get('/', (req, res) => {
 
-    let today = new Date()
-    let options = {
-        weekday: "long",
-        day: "numeric",
-        month: "long"
-    }
-
-    let day = today.toLocaleDateString('en-us', options)
-
     Item.find({}, (err, foundItems) => {
-         res.render('list', { listTitle: day, newListItems: foundItems})
+         res.render('list', { listTitle: "Your List", newListItems: foundItems, day: day})
     })
 })
 
@@ -57,27 +53,66 @@ app.get('/', (req, res) => {
 //Add list items
 app.post('/', (req, res) => {
    const itemName = req.body.newItem
+   const listName = req.body.list
 
    const item = new Item({
        name: itemName
    })
 
-   item.save()
+   if (listName === "Your List"){
+       item.save()
+       res.redirect('/')
+   }else{
+       List.findOne({name: listName}, (err, foundList) => {
+           foundList.items.push(item)
+           foundList.save()
+           res.redirect(`/${listName}`)
+       })
+   }
+})
 
-   res.redirect('/')
+//Custom lists
+app.get('/:customListName', (req, res) => {
+    const customListName = req.params.customListName
+    List.findOne({name: customListName}, function(err, foundList) {
+        if (!err){
+            if (!foundList){
+                const list = new List({
+                    name: customListName,
+                    items: item1
+                })
+            
+                list.save();
+                res.redirect(`/${customListName}`)
+            }else {
+                res.render("list", {listTitle: foundList.name, newListItems: foundList.items, day: day})
+            }
+        }
+    })
+
+
 })
 
 //Delete list items
 app.post('/delete', (req, res) => {
     const checkedItemId = req.body.checkbox
-    
-    Item.findByIdAndDelete(checkedItemId, (err) => {
-        if (!err){
-            console.log('Item deleted')
-        }
-    })
+    const listName = req.body.listName
 
-    res.redirect('/')
+    if (listName === "Your List"){
+        Item.findByIdAndDelete(checkedItemId, (err) => {
+            if (!err){
+                console.log('Item deleted')
+            }
+        })
+        res.redirect('/')
+    }else{
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) => {
+            if (!err){
+                res.redirect(`/${listName}`)
+            }
+        })
+    }
+    
 })
